Bug 692365: Don't derive metrics for CIDFontType 2.
If the metrics don't exist in a CID font, we attempt to derive "sane"
metrics from the bounding box information. This is problematic since a
font (even with WMode 1) missing vertical metrics should be treated as a
horizontal (i.e. WMode 0).

However, we *do* still want to derive vertical metrics, for a vertical
writing font when we're using a Truetype font as a substitute for a
"real" CIDFont.

No cluster differences expected.