Add Memento configuration to Visual Studio projects (and gs makefile)
When swapping between memento and non-memento builds, a complete rebuild
is required due to the pervasive nature of the memento defines. As such
it makes sense to have it as a separate configuration in the project.

Memento directories are named 'memobj', 'memobj64', 'membin' etc in
keeping with 'debugobj', and 'profobj' etc.

Also, take the opportunity to fix a couple of small things in the visual
studio files. Firstly, the 64 bit profile build targets were incorrect
(weren't setting paths etc).

Secondly, we remove the use of nmake /A for the rebuild step, and instead
rely on an explicit clean call, then a rebuild; this means we get
consistent results for a rebuild compared to a clean then a build.

As part of this work, we add logic to the (MSVC) makefiles to detect
MEMENTO=1 and PROFILE=1. If set, these change the default DEBUG, TDEBUG
and DEBUGSYM flags as appropriate, and change the default output
directory names. This keeps the invocations from Visual Studio sane
(as I found lots of mistakes in them as I looked through).

Also fix some places where we were failing to set and pass on
BINDIR/PSOBJDIR resulting in stray directories being created.

CLUSTER_UNTESTED