genarch tweak for cross-compiling; supply ready made arch files.
Previously genarch would unconditionally generate an arch.h file for the
current platform every time it was called. This fails for cross-compiling,
forcing the makefile to be edited every time. In this commit we add a
tweak that solves this problem.

Now, genarch can optionally copy an existing arch.h file; if the make
(or nmake) invocation has TARGET_ARCH_FILE=path/to/existing-arch.h added
to it, then instead of generating an arch.h for the current (host)
platform, it will instead copy the supplied one (assumed to be appropriate
for the target platform).

MSVC builds are updated here to automatically pick a TARGET_ARCH_FILE
based upon the WIN64 define. This should allow us to build 32bit binaries
on a 64bit box from the Visual Studio solution. We still cannot build
64bit binaries on a 32bit box using the unedited solution, as we have
no way to detect the type of box we are on, and hence cannot select
the appropriate AUX compiler to use.

A new directory, 'gs/arch' is added, containing the 2 windows msvc arch.h
files. More files can be added in future, ideally keeping to the
<os>-<architecture>-<toolset>.h naming convention.