Fix for patterns with plank device, transparency, and clist.
It turns out that when we right the tile raster for the clist we don't
really know what the "true" target of the pattern is only the final
target device.  In this particular situation, the target for the pattern
is a pdf14 device, the tile has no transparency, and the final target
is a planar device.   When ended up assuming the tile was planar which
it was not and ended up with a seg fault