Bug 691991: allow extra compile/link flags.
The XCFLAGS and XLDFLAGS should be available to be specified as parameters to
"make", but they were being used by the autoconf tools. This change moves the
autconf set flags into their own variables, leaving XCFLAGS and XLDFLAGS for
use at the make command line. Similarly, it adds an XTRALIBS variable which
can also be set from the make command line.

This is a slight modification of a patch originally from William Bader.

No cluster differences.