Allow planar device to make use of special getbits handling for 1bit cmyk.
We have a special function gx_get_bits_copy_cmyk_1bit, used to quickly
map cmyk (1bpc) to rgb. Our use of this is dependant on the
gxdso_is_std_cmyk_1bit call. Unfortunately this was broken for planar
devices, so plank was never getting the speed benefits of this routine.
Fixed here.