Bug #692515, array index out of bounds (static analysis)
The routine pdf_SepCMYK_ConvertToRGB accessed element 5 of an array
passed as a parameter, when it should have accessed element 4 in order to
convert a CMYK value into an RGB value.

The code path is only exercised when creating PDF/A or PDF/X, the
ProcessColorModel is DeviceRGB, and the input contains a /Separation or
/DeviceN colour space with a DeviceCMYK /Alterenate space.

No differences expected.