Bug #692520 (partial, pdfwrite portion only) fix return testing
The cos_c_string_value function returns a pointer (to a parameter passed to it!)
which is cast to an int and tested for negativity. This is terrible practice.
On 64- bit machines the pointer doesn't fit into an int and on 32-bit machines it
will erroneously detect an error if a valid pointer exceeeds 2GB.

The only reason for the test is to allow the funciton call to be included
in the body of an 'if' test, which is a hideous way of constructing the
code.

This commit reformats the code, and does away with the irrelevant return code
test.

No differences expected, this path is not tested by the cluster.