Add special case for faster planar rops in greyscale and rgb modes.
If we're doing a planar rop, and (S is either unused, or constant) and
(T is either unused, or constant), and (we're in rgb or greyscale), then
we can treat each component separately. This avoids pulling the planes
together, converting to rgb, doing the rop, converting back to cmyk and
splitting back out again.

I had hoped to use a variant of this code to do cmyk rops too, but that
turned out to be impossible; even if we 'rewrite' the rop on entry to
change it from an rgb one to a cmy one, the non-orthogonal nature of the
K plane bites us.

CLUSTER_UNTESTED as not currently enabled on the cluster.