Bug #692488 add limited support for MM fonts to ps2write
ps2write currently emits MM fonts with the WeightVector but no other parts
of the font, and so it is not a valid MM font. However it also emits MM
OtherSubr calls in the CharString, which is not legal and causes errors.

This code parses (but does not interpret) the Subrs and CharStrings and
removes the blending values, leaving only the base values.

This is not a comprehensive solution, the parser can be defeated by a
sufficiently convoluted CharString/Subr combination and in addition the
code should really blend the vlaues instead of discarding them. This would
require writing a new version of 'psf_convert_type1_to_type2' which emits
type 1 CharStrings instead of type 2. The bug has been left open and
converted to an enhancement to reflect this ambition.

This code shows 2 progressions in the regression test suite:
test-setweightvector.ps
fts_20_2005.pdf