Fix for Bug 692494 and Bug 692433.  Overprint issues.
The fix for bug 692494 involved making sure that we use the original floating point values in
determining if we have a nonzero component so that overprint mode occurs properly
for very small values.  Thanks to Ken and Ray for the initial analysis on this bug.

A study of 692433 revealed that AR does not do overprint simulation when the source
ICC profile does not match the destination profile.  To test this, I created a
test file that had 2 overlapping fills with CMYK values that were specified by
SWOP CMYK values.  If the simulation profile in AR matched this profile, then overprinting
was properly displayed by AR.  If not, then overprinting was not displayed.  This
makes sense as the overprinted inks really should be in the color space for the
target device.   Hence, if the source and destination profiles do not match we will
not be showing overprint  This test file is called overprint_icc.pdf and will be added to
our test files.