Partial fix for planar device pattern rop problems.
plank vs pamcmyk4 tests have revealed various files still rendering wrongly
in plank format. These are to do with a halftone being rendered as a pattern
and then used as part of a rop.

Specifically, if the 'rop source' device is used, then it is fed a pattern
tile that is in planar format. When this calls strip_copy_rop, the pattern
tiles data is used as the 'T' field; strip_copy_rop assumes that both T
and S are always in chunky format, and so garbage is read for the T plane.

The fix, implemented here, is to spot that the pattern tile is planar, and
to set a new bit (lop_t_is_planar) in the lop to indicate this fact. The
planar device can then catch this bit and can handle it.

Currently, this is implemented by doing a planar_to_chunky operation into
a temporary buffer, performing the strip_copy_rop, and then freeing the
buffer. Later we may well implement cores that can cope with the common
cases with T being still in planar form.

This commit only fixes unbanded operation; currently when the clist writer
puts the tile into the clist, it has no way of indicating that the tile
is in planar format, hence only writes 1/4 of the data. I need to talk to
Ray and Michael about this. This is broken already, so this commit shouldn't
make anything worse.

I have updated the docs (out of date since 1998, at least) in related areas,
and fixed some broken english.