Add num_planes to gx_strip_bitmap for planar devices.
Here we alter all existing uses (that I can find) to set num_planes
as appropriate.

We update the clist code to encode/decode the new num_planes field. In
order to fit the extra bit required to signal that this is present into
the encoding byte we optimise the representation of depth down to 4 bits
from 5.

We remove the lop_t_is_planar hack as we now have a cleaner scheme for
signalling it.

This should hopefully remove the last differences between pamcmyk4 and
plank.