Bug 692561: fix handling of tr 3 in setshowstate
In setshowstate we have to work around buggy PDFs in a manner which
means we "show" the text in question, but then rely on a charpath
operation to update the current point accurately. The problem was
that the currentpoint had already been updated by the show operation.

This change ensures we return the currentpoint to the correct position
before doing to charpath to get the accurate text advance.

No cluster differences.