Improve PDFFitPage to center the PDF Box on the page.
This was requested by customer 531 and seems reasonable. We didn't document exactly
where the PDF was placed on the page after 'fitting', so centering seems like an
acceptable behavior, so I didn't add a command line option to lower/left position.
If we get any complaints, we can add this.