Remove the 'return' from the macro gx_set_dev_color
Macros shouldn't do a return, its highly surprising to a developer. In addition
I need access to the return value for pdfwrite, and potentially other
devices may do so as well. In this case the return value is Remap_color
which happens when a Pattern is set.

No differences expected in regression tests.