Fix Bug 692569; SEGV in threshold halftoning code.
penum->xci can be negative in clist rendering of landscape cases.
The % operator in C is broken (or at least counter-intuitive) so
the code goes wrong; fix it up with some tweaks to the logic
afterwards.

In a future commit, we'll move from % 16 to & 15 (does the right thing,
and is a lot faster).

No differences expected.