Fix PDFFitPage when PDF orientation doesn't match page orientation.
There was logic to determine if the pdf Box (after /Rotate is applied) orientation
matches the orientation of the PageSize (landscape vs. portrait), but the rotation to
make the image match the Page orientation was missing. Previously this caused the
scaling to be wrong and caused the image to be clipped. Adding the 90 degree rotate
fixes it. Feature needed by customer 531.