Bug 689450: deal with invalid font entry in res dict.
In the event that an entry in a resource dictionary for a font is or
references an invalid object type (i.e. not a dictionary), issue a
warning and use the same logic as we do for a missing font resource.

For this error, we'll also honor the PDFSTOPONERROR setting.

In addition, as I was touching the logic around the same place, this
commit changes the missing font resource case to also honor PDFSTOPONERROR.

No cluster differences expected.