Bug 692534: revert to unit matrix for Microtype fonts.
I changed the MT fonts to use a "conventional" Type 1 font matrix to
make a Quality Logic file "work" as expected - it's a daft test which
makes the invalid assumption that the fonts in the standard font set
will *always* use a 1000x1000 design matrix. That assumption is invalid
since nothing in the spec dictates that the standard fonts must be Type 1
(clearly as in this case we're using MT fonts!), nor even that Type 1
fonts have to use a 1000x1000 font matrix.

As the change breaks a valid file, to try to satisfy a file which relies
on an invalid assumption, I'm reverting it.

CLUSTER_UNTESTED.