Better preservation of PDF text rendering modes in pdfwrite.
Bug #688267, #692243. The 'swapcolors' is now moved into pdfwrite, instead
of being performed in PostScript in the PDF interpreter. The 'spaced'
variants of show now perform simlar techniques to the 'pdfwrite' text rendering
routines, when teh device supports text rendering modes.

Caveats: pdfwrite always emits text enclosed in gsave/grestore. Because of
this we cannot preserve any of the text rendering modes involoving clipping
as the grestore also restores the clip path! So we subtract 4 from the mode
and emit the text that way, then handle the clip separately.

Because text_process doesn't expect to receive gs_error_Remap_Color errors
(which cause the interpreter to run Pattern PaintProcs usually) we can't
set the stroke colour during text processing (which is how it worked before).
Instead we set the colours during text_begin. We don't actually write the colours
to the PDF file at that point though, because that causes problems synchronising
graphics states. Instead we leave the emission of the colour unchanged, we
just evaluate the colours in text_begin.

There is some weirdness in the PDF interpreter which I do not understand.
Most cases are surrounded by 'currentlinewidth exch.......setlinewidth'
which preserves the current line width in case we have to change it for
stroked text. In one case, however, this causes files to fail with an error.

I have tried without success to unravel the PDF interpreter to figure out
what is going on. Since I can't work it out I have created a dictionary,
stored the linewidth in that, and then pulled it back out and restored at
the end. I did try wrapping a gsave/grestore round the operation (taking
care to preserve the modfified currentpoint) but that caused even more
problems. Again I have no idea why.

I would like Alex to look into this so I'm leaving one of the bugs open
and re-assigning to him. Also he will probably want to reformat the code
I've added to the PDF interpreter.