Fix longstanding color halftone limitation.
gx_dc_ht_colored_fill_rectangle in gxcht.c contains a long standing
limitation (bug/defect), in that when doing strip_copy_rop with
subrectangles, it fails to offset the source samples for each call.
This is noted in the form of a comment.

This is fixed here.

There appears to be an underlying bug in this code (or in the code it
calls) in that the 'optimisation' case (do it all in a single
tile, rather than subdividing) gives different results to the normal
(subdividing) case. See bug 692609.

This can be shown with tests_private/pcl/pcl5cats/Subset/PWEP6SC1.BIN
at 300dpi with or without banding to the plank or pamcmyk4 device.

The limitation that this patch fixes does NOT solve the problem - but
I fixed it in the hopes that it would. Committing the code here as it's
an improvement anyway.