Memento buld tweak; valgrind warnings quieted for temporary buffers.
Temporary buffers are used in the planar device and the color halftoning.
These are uninitialised by default, which means that the last 'part byte'
transferred is thought of as being undefined in valgrind.

In memento builds we memset them to 0 to avoid this.

CLUSTER_UNTESTED as memento builds are not testing in the cluster.