Fix enhancement bug 692615: Create a platform specific hook to set the serialnumber
The main change is to add the hook, but for common platforms, try to get a reasonably
unique serialnumber based on the hardware (we hope). Initial platform in this patch is
only Windows. We would like to support linux/unix and mac OS/X. If the mechanism fails,
just return gs_serialnumber set from GS_SERIALNUMBER in gscdefs.c