Font copying does not work with CIDFonts using dictioanry form of GlyphDirectory
Bug #692573, the font copying code uses 'slots' to record glyph data, which
are entries in an array. The index into the array can be a name or a CID.

This assumes that CIDCount (the number of glyphs in the font) is the same as
the maximum CID (ie the range is continuous), which is generally true.
However, the dictionary form of GlyphDirectory allows for discontinuous CIDs.
This meant that the array was too small to use the CID as an index.

Fixing this efficeintly would mean completely rewriting the code in gxfcopy.c
so that it uses some kind of dictionary structure instead of an array. Instead
I've chosen to find the highest CID in the font (by enumerating the GlyphDirectory)
and use that to create he array with enough entries. Inefficient use of memory
but it saves having to re-architect the code.

The text that was previously missing is now present, and this reveals a new
problem, some of the text is incorrectly spaced after conversion with pdfwrite.
I'll open this as a new bug.

No differences expected