Fix to handle tiff associated alpha images correctly in XPS
TIFF images in XPS can come with an alpha channel that may or may
not be premultiplied.  In our existing XPS flow, we currently store
the alpha channel in a separate channel and treat it as an image mask
in the graphics library.  This is fine, but if the image data was
premultiplied by the alpha data, we will need to undo this operation
otherwise we end up applying the mask on image data that is already
scaled by the alpha data.