Fix for Bug 692505  XPS patterns with transparency.
There were essentially three issues two in the xps interpreter and one in the graphics library.

First in the xps interpreter we need to know if the pattern object has transparency and to
set up the pattern object for the graphics library so that the graphics library will
know to push the pdf14 device on the pattern accumulator object.

We also had the issue that we were not undoing the pre-multiplied alpha in the tiff image
in the interpreter.

Second issue is that in ghostscript, we call pattern_paint_finish from
the pdf interpreter when we are all done setting things up for the pattern.  This is now
handled prior to the call before gx_pattern_cache_add_entry which is where it is done in
pattern_paint_finish.

Also, some debugging of the mask stack was added in gdevp14.c.  This was needed to track down
a bug that become visible when I added in the pattern transparency support.  The issue
was that the parent PDF14 device may often have a valid soft mask in its mask stack if there
had been a previous soft mask push.  Since we made the soft mask state be part of the q Q
operations in the PDF interpreter (the soft mask is actually part of the extended graphics state
in PDF), we have to make sure to pop the soft mask we we are all
done using it in the XPS interpreter.  I added the appropriate commands for that.

There was also an issue that cropped up with the pdf14 device and the interaction of the image
interpolation code.   We always interpolate image in XPS and it turned out that if we had an
opacity mask pattern created with an image we installed a gray pdf14 device for the pattern.
The interpolation code was incorrectly using halftoning for the pdf14 gray device due to the
color info settings in that device.

This commit results in many progressions in PDF and XPS (450 differences).  In PDF, if the
target was a gray devices and the content had
interpolated images with transparency we see progressions due to the fact that we were improperly
halftoning the images when the trans device is a contone device.  In XPS, several missing
figures now are visible and we correctly handle tiling with objects that have transparency.