Fix bug 692618. Crash and Object not in any chunk error and VMerror.
Actually two problems showed up after the previous commit related to this bug. A crash
could occur (or cause "Object not in any chunk") because the pdf14 device did not use
stable_memory for separation names. The second issue was a VMerror when the decision
for banded (clist) mode was flawed, because if the device supported separations the
estimates for transparency buffers was too small because the call to
gdev_prn_allocate_memory was sometimes called with new_width, hew_height = 0,0 which
expected the 0,0 to be ignored. The new_width and new_height are only used from
gdev_prn_reallocate_memory.

The non-encodable pixel issue still remains.