Add UTF-8 as an output option to the txtwrite device.
The TextFormat parameter now has three values; 0, 1 or 2 to allow for 'XML'
output with full information, simple UCS2 or simple UTF-8 output.

Altered the meanings of the TextFormat so that 0 is now the 'XML' output
which allows a more sensible selection of 1 for UCS2 or 2 for UTF-8, since
these output foramts are otherwise identical. Modified the default to be
'2' (UTF-8).

Updated the documentation in devices.htm.

Also removed a number of C++ comments.

No differences expected, this device is not cluster tested.