Add new 'long form' debugging flags.
Add a table of 'known' debugging flags to ghostscript (gdbflags.h). These
list all the currently defined -Z flags, together with 'long form' names
for them (i.e. --debug=tiling will do the same as -Zt).

Add a new routine to list the known flags (and hook this up so that
--debug on the command line lists the flags).

Add a new routine to parse a set of flags passed to --define= on the
command line (and hook that up).

The code that does:

 if_debug6('c',...)

all remains unchanged, but we now have the ability to add new flags,
and to do:

 if_debug6(gs_debug_flag_whatever, ...)

in the same way. This frees us from the limitations of only having
a limited number of easily representable 'character' flags.