Bug 692605: resolve FAPI/UFST matrix clash
The Microtype fonts require a unit design grid, but many jobs assume that the
standard fonts set have a design grid of 1000x1000. For those two conflicting
requirements to be satisfied, the UFST integration layer has to compensate
for the scaling.

As usual with the UFST code, that exposed a host of other problems, including
failing to produce a bitmap sometimes when one is require for the glyph cache
and failing to handle a glyph too big for UFST to image (i.e. fallback to an
outline) in some cases.

No cluster differences expected.