Make txtwrite 'XML' output compatible with MuPDF output.
The txtwrite device now outputs 'XML' (something like XML) in a compatible
fashion to the '-ttt' output from MuPDF. It does not yet attempt to create
blocks of text, it just outputs the data it received as it received it,
there is no ordering of the data whatsoever. The format is compatible with
the MuPDF format though.

No differences expected, txtwrie is not cluster tested.