Fix stroke_adjust logic; apply adjust in fewer cases.
In 'adjust_stroke', we decide whether to apply stroke adjustment or not.
With the logic as it stands, the only time we DO NOT apply stroke
adjustment is when stroke adjustment is disabled, AND we have a diagonal
stroke segment.

My understanding of stroke adjustment is that it is only supposed to
apply when we have a horizontal or vertical line, AND when it is
configured on. In line with this, I have changed the logic so that
we DO NOT apply stroke adjustment either when stroke adjustment is
disable, OR we have a diagonal stroke segment.

This produces many thousands of differences in the cluster testing, but
having looked through the first bmpcmp load of them (1000 files), they
all look to be exactly as expected with no changes so horrific as to
stop this commit.