Fix for (part of) bug 692513.
When drawing the a radial shading, we approximate the join between the
two circles by 4 tensor patches based on joining the 4 quadrants of
the circles in question. The order in which these were being drawn was
causing unexpected overwriting of pixels.

Here I have rejigged the order in which they are drawn to solve the
problem.

Cluster testing shows various progressions, and only one regression.
This has been noted as Bug 692657, and shows signs of it being an
independent bug.