Fix for bug 692639.  PDF14 mono device initialization.
With the fix of the max_gray initialization in pdf14 device to avoid halftoning during interpolation of images (see description in 2b91a85069d421465220a83c82ea491733d15017)
a conflict occurred when checking if the device was separable and
linear.  Setting of the dither_grays for the pdf14 gray device
fixes this  issue.