Further enhance txtwrite output; detect blocks of text.
Yet another output mode for txtwrite will now attempt to detect blocks of
text and output the blocks. This is similar to the MuPDF '-ttt' output
but, since the algorithm is entirely different, the output will not be
identical.

Updated the documentation in devices.htm.

No differences expected, txtwrite is not regression tested.