Bug 692646: remove bad call to memcmp()
The memcmp() to compare the matrices (to decide if we need to rescale the font)
was the address of the pointer to the matrix, rather than the pointer itself.

This change addresses that by changing the memcmp() to an explicit check for
the equality of each element of the arrays. The reason I've opted for this is
because, for example, [1 -0 0 1 0 0] == [1 0 0 1 0 0] which an explicit
equality check will evaluate correctly, but a memcmp() will fail the check.

This causes what seems to be a progression in comparefiles/Bug689006.pdf.
Other than that, no cluster differences are expected.