Bug 692470: remove sscanf() in CFF code
On Linux (at least) the locale affects the behaviour of sscanf(): certain
languages use a comma (",") for the decimal point, rather than a full stop
("."). Setting one of those locales (such as Spanish) causes sscanf() to fail
when parsing a number with a fractional component in the CFF interpreter.

Replace the sscanf() call in the CFF interpreter with a call to Ghostscript
PS interpreter's internal scan_number() function which is unaffected by locale.

No cluster differences.