Fix bug 692623. clist_get_data and clist_put_data changed to use int64_t offset.
The above functions are used by gx_dc_pattern_read_raster and gx_dc_pattern_write_raster
which are instances of the dev_color_proc_read and dev_color_proc_write procs of the
gx_device_color_type_s struct. The definitions for all color types is changed to use
int64_t (as well as the prototype) and the stdint_.h is added to gxdcolor.h.