pdfwrite - Improve font naming when standard fonts are not embedded.
Bug #692608; When the input is PCL, 'standard' font names are used (eg Courier),
and font embedding is false, the output PDF file contains the names of the
URW fonts which are used as substittues for the standard fonts.

This means that the resulting PDF needs to be read on a system with URW fonts
installed, or the fonts will be substituted, which may result in poor choices.

By instead embedding the correct original (standard) font name we can get
better substitution.

Thanks to Hin-Tak Leung for the original work and the patch which has been
adopted with slight modifications to pacify compilers.

No diffreences expected.