Fix type 1 font copying code SEAC scanner.
When embedding type 1/CFF fotns the code scans the glyphs to see is any of
them are SEAC (Single Encoding Accented Characters) glyphs and expands
them if they are.

The CFF parsing was skipping over the 'shotint' operator (itself an awful hack)
instead of pushing the value onto the stack. If the shortint was the index
for a /Subr this led to the wrong Subr being called. Because Subrs expect
parameters on the stack, calling the wrong one leads to stack cuorruption
and in this case caused a crash.

No differences expected.