Properly finalize the DCT-writing memory device to avoid memory leaks
Bug #692670 "Ghostscript uses 2 gigs of memory reading file". When writing
images a JPEG the pdfwrite code creates a memory device temporarily. It
was not, however, finalizing the device. Previously this did not cause a
problem, but with the ICC colour work, this left the ICC buffers set up,
causing a ~6Kb per image memory leak.

This commit explicitly finalizes the device, which frees the memory and
gets rid of the leak.

No differences expected