ps2write - improve TrueType composite glyph handling
Bug #692626, #696264.

When emitting a TrueType font from a CIDFont with TrueType outlines if a
glyph was a composite glyph we did not create an entry in name table for
any components of the glyph which were not already present in the output
font.

This caused problems creating the CMAP subtable for the embedded TrueType
font.

This patch shows progressions with ps2write and files CATX4638.pdf, bug690450.pdf.
A regresison with pdfwrite and file bug688421.pdf and a difference in
Bug 691121.pdf with pswrite (wrong before, differnetly wrong now)

Despite the regression I'm committing it, as it shows significant improvements
for ps2write.