Addition of thresholding functions for subtractive color spaces.
Previously, I had swapped the threshold and the image data to handle the subtractive case.
Unfortunately due to the way things were written that caused issues.  So for now we added
unique functions for the subtractive case to avoid an additional conditional in the low
level part of the code.