Clarification of the gen_ordered settings and README.
The current default value for -s (size of the super cell) is 1
which indicates that them smallest possible size should be used.
This occurs regardless of the requested quantization levels.
This is clarified in the README and a message is displayed
during the creation of the screen providing a minimum suggested
value for -s to achieve the number of requested levels.