Bug 692683: revise FAPI/FT code so it uses the GS file handling
When the Freetype interface was called to create a font object with a path
to the font file, it previously simply passed the path into Freetype, and
left it to handle the file details.

This revision changes that so that we build a custom Freetype stream object
which uses the Ghostscript file API to access the data. We generally prefer
to operate that way anyway, and (more importantly) this allows FT "direct"
acccess to font files stored in our ROM filesystem.

No cluster differences.