Bug 692672: slightly revise symbolic fonts in PDFs
This is a slightly better of approximation of what Acrobat does in the out
of spec condition when a font has both an encoding in the font object and
the symbolic flag set in the descriptor's flags entry.

If the flags say symbolic, *and* the font contains a symbol cmap table,
ignore Encoding if there is one, treat as symbolic.

If the flags say symbolic, and the font doesn't have a symbol cmap, but we do
have an Encoding, treat it as non-symbolic if we don't have an Encoding, try
the MacRoman cmap table.

Finally, if none of that works, remove the Encoding, and treat it as
non-symbolic. This part is the biggest guess at Acrobat's behavior.

The various tests used in investigating what Acrobat does are attached to the
bug (692672).

No cluster differences.