Bug 692666: grid fit images.
The file in bug 692666 draws many thin horizontal rectangles, then draws
1 pixel high images on top of each one. Current gs 'thickens' the
rectangles (due to the 'any part of a pixel' rule), and they overwrite
the previously drawn image - hence it appears black.

The problem is thus that the images are not similarly stretched.

Images are not supposed to use the "any part of a pixel" rule, where
vectors are, but the fact this file works in acrobat would seem to
contradict that. Instead we try a simple fix; if we are using the
'any part of a pixel' file (i.e. fill_adjust != 0), and we meet an
ImageMask image that is either 1 pixel high, or 1 pixel wide, then we
ensure that it is 'stretched' to completely fill any pixels that it
touches.

We specifically do NOT perform this stretching if we are in the process
of rendering a glyph.

This produces various rendering diffs, none that offensive.