ps2write - only execute setpagedevice if /PageSize changes
Bug #692691

The ps2write output (being based on pdfwrite) always emits a /MediaBox as
this is a required key for PDF. Previously the PostScript prolog would
always emit a setpagedevice using the MediaBox as the PageSize entry.

This caused a problem for CUPS as it inserts PPD code into the ps2write
output, and if this enables Duplex then each invocation of setpagedevice
will flush the accumulated output, which defeats Duplex.

The prolog code now tests the current page size and only executes
setpagedevice if the newly requested size differs from the current size. In
a simple attempt to ignore small differences we round the current and requested
media sizes, and convert to integers before comparison.

No differences expected in regression tests