Bug 692681: handle invalid Encoding arrays in Type 3 fonts
Distiller quietly converts the contents the contents of a Type 3 font's encoding
array to names (at least if they are integer objects), whilst CPSI throws an
error.

We will now do as Distiller and convert to name objects, unless we are in "CPSI
compatibility mode", when we will throw an error.

No cluster differences.