Correct CIDSet and CIDToGIDMap generation when LastChar > number of glyphs
More CIDSet work. WHen LastChar is > number of glyphs in font (already a subset)
we were setting the length of the CIDToGIDMap and the CIDSet to the LastChar
value. This should be LastChar + 1 to allow for the /.notdef.

No differences expected, this only happens when generating PDF/A which is not tested.