Add a "dummy" call to validate UFST font.
It turns out that UFST's CGIFfont() call does little or no validation of the
font for which it is creating an object. So, add a "dummy" call to
CGIFfont_metrics() which will return an error if, for example, a request
for a font index which doesn't exist in the current FCO is made.

CLUSTER_UNTESTED